% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuals.R
\name{getAllPlayerPoints}
\alias{getAllPlayerPoints}
\title{Get All Player Fantasy Points}
\usage{
getAllPlayerPoints(players_link, week, statType = 2, sortType = 2,
  FA_only = TRUE, position = "ALL")
}
\arguments{
\item{week}{Week number}

\item{teams}{Data frame of teams from getTeams function}
}
\value{
Each team's roster with fantasy points and starter designation.
}
\description{
Get All Player Fantasy Points
}
\examples{
getPlayerPoints(teams, 1)
}
