% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getLinks}
\alias{getLinks}
\title{Get links from Fleaflicker website}
\usage{
getLinks(leagueID)
}
\arguments{
\item{leagueID}{Your Fleaflicker league ID}
}
\value{
Links added to your environment
}
\description{
Get links from Fleaflicker website
}
\examples{
getLinks(123456)
}
