% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{getPlayerProjections}
\alias{getPlayerProjections}
\title{Get Projected Player Scores}
\usage{
getPlayerProjections(players_link, week, statType = 7, sortType = 7,
  FA_only = TRUE, position = "ALL")
}
\arguments{
\item{week}{Week Number}

\item{statType}{The statistic that is shown on the page - defaults to projection.}

\item{sortType}{The statistic to sort on - defaults to projection}

\item{FA_only}{Boolean variable that determines if only free agents are included}

\item{position}{Player position to get projections for - defaults to ALL}

\item{plaers_link}{Web link to main league page}
}
\value{
Teams and projected total score.
}
\description{
Get Projected Player Scores
}
\examples{
getPlayerProjections(players_link, 1, 7, 7, TRUE, 'QB')
getPlayerProjections(players_link, 1, 7, 7, FALSE, c('QB','K'))
}
