% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuals.R
\name{getRosteredPlayerPoints}
\alias{getRosteredPlayerPoints}
\title{Get Rostered Player Fantasy Points}
\usage{
getRosteredPlayerPoints(rosters, week)
}
\arguments{
\item{week}{Week number}

\item{teams}{Data frame of teams from getTeams function}
}
\value{
Each team's roster with fantasy points and starter designation.
}
\description{
Get Rostered Player Fantasy Points
}
\examples{
getPlayerPoints(teams, 1)
}
