% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{getTeamProjections}
\alias{getTeamProjections}
\title{Get Projected Team Scores}
\usage{
getTeamProjections(league_link, scores_link, week)
}
\arguments{
\item{league_link}{Web link to main league page}

\item{scores_link}{Web link to "scores" page.}

\item{week}{Week Number}
}
\value{
Teams and projected total score.
}
\description{
Get Projected Team Scores
}
\examples{
getTeamProjections(league_link, scores_link, 1)
}
