% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getTeams}
\alias{getTeams}
\title{Get teams from main league page}
\usage{
getTeams(link)
}
\arguments{
\item{link}{Link to main league page}
}
\value{
Data frame with team ID and team name
}
\description{
Get teams from main league page
}
\examples{
getTeams(123456)
}
